# Change StringView constructor to take signed length.