Currently, It is possible to pass a negative size to StringView which due to implict casting creates bad values. We change signature to ensure we dont get passed bad values. 
Fixes #4331. 

Note, that we do not add this check for the user defined literal _sv since it's guaranteed to be non negative (https://en.cppreference.com/w/cpp/language/user_literal) .