@kgpai 2023-03-23
Here are results of benchmark, they are almost similar : 

```
With SV changes:

/Users/kpai/src/Velox/cmake-build-release/velox/type/tests/velox_string_view_benchmark
============================================================================
[...]ox/type/tests/StringViewBenchmark.cpp     relative  time/iter   iters/s
============================================================================
runStringViewCreate(INLINE_SIZE)                           11.14ns    89.76M
runStringViewCreate(NON_INLINE_SIZE)                        8.24ns   121.36M

without VELOX_CHECK in StringView:

/Users/kpai/src/Velox/cmake-build-release/velox/type/tests/velox_string_view_benchmark
============================================================================
[...]ox/type/tests/StringViewBenchmark.cpp     relative  time/iter   iters/s
============================================================================
runStringViewCreate(INLINE_SIZE)                           11.13ns    89.86M
runStringViewCreate(NON_INLINE_SIZE)                        8.24ns   121.36M
```