@kgpai 2023-03-23
Validated that the call to StringViewConstructor is inlined.